requirements.txt
alembic==1.13.2
annotated-types==0.7.0
blinker==1.8.2
certifi==2024.6.2
click==8.1.7
distlib==0.3.8
filelock==3.15.4
Flask==3.0.3
Flask-Alembic==3.0.1
Flask-SQLAlchemy==3.1.1
greenlet==3.0.3
importlib_metadata==8.0.0
importlib_resources==6.4.0
itsdangerous==2.2.0
Jinja2==3.1.4
Mako==1.3.5
MarkupSafe==2.1.5
numpy==1.24.4
pandas==2.0.3
pipenv==2024.0.1
platformdirs==4.2.2
pydantic==2.8.0
pydantic_core==2.20.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.1
six==1.16.0
SQLAlchemy==2.0.31
typing_extensions==4.12.2
tzdata==2024.1
virtualenv==20.26.3
Werkzeug==3.0.3
zipp==3.19.2
